RF : This is another way to calculate AUC.
This should be more robust. The idea is that we use all the parameters of the
model, except for the parameter that controls the amplitude of the 'hump' to
calculate what the effects of offset and drift are. The difference is then just
due to the AUC of the 'hump' itself. Thanks to Dora for suggesting this.